refactor(fs): extract native watch utilities
