+++ packages/shared/src/fs/index.ts
@@ -1,5 +1,6 @@
 export { FileWatchService, type FileWatchServiceOptions } from './file-watch-service';
 export { FsService } from './fs-service';
+export { realpathOrResolve } from './paths';
 export type {
   FileReadResult,
   FileStat,
